pyproject.toml
[project]
name = "funsearch-cvrp"
dynamic = ["version"]
description = "Sample-Efficient FunSearch for CVRP using LLM-generated heuristics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-docx>=1.1.0",
    "matplotlib>=3.7.0",
    "python-dotenv>=1.2.2",
    "absl-py>=2.4.0",
    "scipy>=1.17.1",
    "vrplib>=2.2.0",
]
authors = [
    {name = "LAI Yikai", email = "laiyk5@163.com"},
    {name = "QING Ziheng"},
    {name = "QING Yuancheng"},
]
classifiers = ["Private :: Do Not Upload"]

# [project.optional-dependencies]

[project.urls]
Homepage = "https://github.com/laiyk5/funsearch_cvrp"
Repository = "https://github.com/laiyk5/funsearch_cvrp"
Issues = "https://github.com/laiyk5/funsearch_cvrp/issues"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/funsearch_cvrp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.version]
path = "src/funsearch_cvrp/__about__.py"

[tool.uv]
python-preference = "only-managed"
package = true

[tool.uv.pip]
reinstall-package = ["funsearch-cvrp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "hatch>=1.16.5",
    "pytest>=7.0.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.1.0",
]

doc = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
]
